# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_Cell_": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Cell"
          },
          "notice": {
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "notice",
          "data"
        ],
        "title": "Envelope[Cell]",
        "type": "object"
      },
...
      "Envelope_list_Alarm__": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Alarm"
            },
            "title": "Data",
            "type": "array"
          },
          "notice": {
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "notice",
          "data"
        ],
        "title": "Envelope[list[Alarm]]",
        "type": "object"
      },
...
      "Envelope_list_Cell__": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Cell"
            },
            "title": "Data",
            "type": "array"
          },
          "notice": {
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "notice",
          "data"
        ],
        "title": "Envelope[list[Cell]]",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Every error response.",
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "notice": {
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "notice",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "KPIs, configuration, alarms, topology, planning and what-if of a synthetic 4G and 2G network. synthetic data: a generated network, not a real operator.",
    "title": "ran-lakehouse API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/alarms": {
      "get": {
        "operationId": "alarms_v1_alarms_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cell",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cell"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Alarm__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "Alarm notifications in a time range"
      }
    },
    "/v1/cells": {
      "get": {
        "operationId": "cells_v1_cells_get",
        "parameters": [
          {
            "in": "query",
            "name": "technology",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "LTE",
                    "GSM"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Technology"
            }
          },
          {
            "in": "query",
            "name": "vendor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Cell__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "Cells with vendor, technology, band, N_RB and DN"
      }
    },
    "/v1/cells/{cell_name}": {
      "get": {
        "operationId": "cell_v1_cells__cell_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "cell_name",
            "required": true,
            "schema": {
              "title": "Cell Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Cell_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "One cell"
      }
    },
    "/v1/clock": {
      "get": {
        "operationId": "clock_v1_clock_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ClockStatus_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "Clock status: the live run and how far each layer has got"
      }
    },
    "/v1/cm/changes": {
      "get": {
        "operationId": "cm_changes_v1_cm_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cell",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cell"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_CmChange__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "CM change log in a time range"
      }
    },
    "/v1/cm/snapshot": {
      "get": {
        "operationId": "cm_snapshot_v1_cm_snapshot_get",
        "parameters": [
          {
            "in": "query",
            "name": "cell",
            "required": true,
            "schema": {
              "title": "Cell",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "At",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_CmObject__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request outside what one answer serves"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such cell, KPI or value"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or out-of-bound request"
          }
        },
        "summary": "One cell's CM (and its relations) at a time"
      }
    },
...
  }
}